Optimize GPU inference by using bfloat16 instead of float32
Changed torch_dtype from float32 to bfloat16 for 2-4x faster inference on GPU. This should reduce per-sample time from ~3 minutes to under 30 seconds.

+++ scripts/collect_stage2_steering.py
@@ -57,7 +57,7 @@ if tokenizer.pad_token is None:
 model = AutoModelForCausalLM.from_pretrained(
     MODEL_NAME,
     device_map="auto",  # Use GPU if available, else CPU
-    torch_dtype=torch.float32,
+    torch_dtype=torch.bfloat16,  # Much faster on GPU
     trust_remote_code=True,
 )
